+++ src/resources/views/bundles/edit.blade.php
@@ -216,4 +216,5 @@
         }(window.jQuery);
     </script>
     @include('redminportal::plugins/tinymce')
+    @include('redminportal::plugins/tagsinput')
 @stop