+++ src/resources/views/modules/create.blade.php
@@ -189,4 +189,5 @@
         }(window.jQuery);
     </script>
     @include('redminportal::plugins/tinymce')
+    @include('redminportal::plugins/tagsinput')
 @stop