+++ src/resources/views/modules/edit.blade.php
@@ -231,4 +231,5 @@
         }(window.jQuery);
     </script>
     @include('redminportal::plugins/tinymce')
+    @include('redminportal::plugins/tagsinput')
 @stop