+++ src/resources/views/products/edit-variant.blade.php
@@ -219,4 +219,5 @@
         }(window.jQuery);
     </script>
     @include('redminportal::plugins/tinymce')
+    @include('redminportal::plugins/tagsinput')
 @stop