+++ src/resources/views/products/edit.blade.php
@@ -383,4 +383,5 @@
         }(window.jQuery);
     </script>
     @include('redminportal::plugins/tinymce')
+    @include('redminportal::plugins/tagsinput')
 @stop